+++ test/td_cache/domain_cache_test.exs
@@ -19,6 +19,7 @@ defmodule TdCache.DomainCacheTest do
       id: System.unique_integer([:positive]),
       external_id: "domain",
       name: "child",
+      parent_id: parent.id,
       parent_ids: [parent.id],
       updated_at: DateTime.utc_now(),
       descendent_ids: [1, 2]
@@ -40,7 +41,7 @@ defmodule TdCache.DomainCacheTest do
       domain: %{id: id} = domain,
       parent: %{id: parent_id}
     } do
-      {:ok, [5, 1, 1, 1, 0, 0]} = DomainCache.put(domain)
+      {:ok, [5, 1, 1, 1, 1, 0, 0]} = DomainCache.put(domain)
       {:ok, d} = DomainCache.get(id)
       assert not is_nil(d)
       assert d.id == id
@@ -52,12 +53,12 @@ defmodule TdCache.DomainCacheTest do
     end
 
     test "updates a domain entry only if changed", %{domain: domain} do
-      assert {:ok, [5, 1, 1, 1, 0, 0]} = DomainCache.put(domain)
+      assert {:ok, [5, 1, 1, 1, 1, 0, 0]} = DomainCache.put(domain)
       assert {:ok, []} = DomainCache.put(domain)
 
       updated = %{domain | name: "updated name", updated_at: DateTime.utc_now()}
 
-      assert {:ok, [0, 0, 0, 0, 0, 0]} = DomainCache.put(updated)
+      assert {:ok, [0, 0, 0, 0, 0, 0, 0]} = DomainCache.put(updated)
       assert {:ok, %{name: "updated name"}} = DomainCache.get(domain.id)
 
       assert {:ok, events} = Stream.read(:redix, ["domain:events"], transform: true)
@@ -66,14 +67,14 @@ defmodule TdCache.DomainCacheTest do
 
     test "deletes an entry in redis", %{domain: %{id: id} = domain} do
       {:ok, _} = DomainCache.put(domain)
-      {:ok, [1, 1, 1, 1, 0, 1]} = DomainCache.delete(id)
+      {:ok, [1, 1, 1, 1, 1, 0, 1]} = DomainCache.delete(id)
       assert {:ok, nil} == DomainCache.get(id)
     end
 
     test "get deleted domain ids", %{domain: %{id: id} = domain} do
       {:ok, _} = DomainCache.put(domain)
       assert {:ok, []} = DomainCache.deleted_domains()
-      {:ok, [1, 1, 1, 1, 0, 1]} = DomainCache.delete(id)
+      {:ok, [1, 1, 1, 1, 1, 0, 1]} = DomainCache.delete(id)
       assert {:ok, [^id]} = DomainCache.deleted_domains()
     end
   end
@@ -88,6 +89,18 @@ defmodule TdCache.DomainCacheTest do
 
     test "returns error if no matching domain exists" do
       assert DomainCache.external_id_to_id("foobarbaz") == :error
+    end
+  end
+
+  describe "id_to_parent_id_map/0" do
+    test "returns a map with domain_id keys and parent_id values", %{
+      domain: %{id: domain_id, parent_id: parent_id} = domain
+    } do
+      assert DomainCache.id_to_parent_id_map() == {:ok, %{}}
+
+      DomainCache.put(domain)
+
+      assert DomainCache.id_to_parent_id_map() == {:ok, %{domain_id => parent_id}}
     end
   end
 end